+++ store/primary/src/admin/data.rs
@@ -1,9 +1,10 @@
+use replisdk::platform::models::ClusterDiscovery;
+
 use replicante_models_core::actions::node::Action;
 use replicante_models_core::agent::Agent;
 use replicante_models_core::agent::AgentInfo;
 use replicante_models_core::agent::Node;
 use replicante_models_core::agent::Shard;
-use replicante_models_core::cluster::discovery::ClusterDiscovery;
 use replicante_models_core::cluster::ClusterMeta;
 
 use crate::backend::DataImpl;