+++ store/primary/src/backend/mongo/data.rs
@@ -1,6 +1,7 @@
 use failure::Fail;
 use failure::ResultExt;
 use mongodb::sync::Client;
+use replisdk::platform::models::ClusterDiscovery;
 
 use replicante_externals_mongodb::operations::scan_collection;
 use replicante_models_core::actions::node::Action;
@@ -8,7 +9,6 @@ use replicante_models_core::agent::Agent;
 use replicante_models_core::agent::AgentInfo;
 use replicante_models_core::agent::Node;
 use replicante_models_core::agent::Shard;
-use replicante_models_core::cluster::discovery::ClusterDiscovery;
 use replicante_models_core::cluster::ClusterMeta;
 
 use super::super::DataInterface;