+++ devtools/replidev/CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 ### Changed
+- **BREAKING**: Migrate `ClusterDiscovery` to `replisdk` provided model.
 - Update dependencies.
 
 ## 0.2.0 - 2022-09-13